feat: make kanban-board a server component
